requirements.txt
openai
pyautogui
pyperclip
pyttsx3
SpeechRecognition
python-telegram-bot
pystray
Pillow
mss
psutil
pygetwindow
pycaw
comtypes
